requirements.txt
flask==3.1.1
python-telegram-bot==13.15
ccxt==4.1.84
openai==1.30.3
python-dotenv
